fix: fix editor modal
