+++ src/features/Conversation/Messages/AssistantGroup/components/EditState.tsx
@@ -1,7 +1,6 @@
-import { Flexbox } from '@lobehub/ui';
-import { MessageInput } from '@lobehub/ui/chat';
-import { memo, useMemo } from 'react';
-import { useTranslation } from 'react-i18next';
+import { memo } from 'react';
+
+import { EditorModal } from '@/features/EditorModal';
 
 import { useConversationStore } from '../../../store';
 
@@ -11,40 +10,24 @@ export interface EditStateProps {
 }
 
 const EditState = memo<EditStateProps>(({ id, content }) => {
-  const { t } = useTranslation('common');
-
-  const text = useMemo(
-    () => ({
-      cancel: t('cancel'),
-      confirm: t('ok'),
-      edit: t('edit'),
-    }),
-    [],
-  );
-
   const [toggleMessageEditing, updateMessageContent] = useConversationStore((s) => [
     s.toggleMessageEditing,
     s.modifyMessageContent,
   ]);
 
-  const onEditingChange = (value: string) => {
-    updateMessageContent(id, value);
-    toggleMessageEditing(id, false);
-  };
-
   return (
-    <Flexbox paddingBlock={'0 8px'}>
-      <MessageInput
-        defaultValue={content ? String(content) : ''}
-        editButtonSize={'small'}
-        onCancel={() => {
-          toggleMessageEditing(id, false);
-        }}
-        onConfirm={onEditingChange}
-        text={text}
-        variant={'outlined'}
-      />
-    </Flexbox>
+    <EditorModal
+      onCancel={() => {
+        toggleMessageEditing(id, false);
+      }}
+      onConfirm={async (value) => {
+        if (!id) return;
+        await updateMessageContent(id, value);
+        toggleMessageEditing(id, false);
+      }}
+      open={!!id}
+      value={content ? String(content) : ''}
+    />
   );
 });
