+++ src/features/ModelSwitchPanel/utils.ts
@@ -1,0 +1,1 @@
+export const menuKey = (provider: string, model: string) => `${provider}-${model}`;
